+++ migration/src/lib.rs
@@ -1,6 +1,5 @@
 pub use sea_orm_migration::prelude::*;
 
-mod m20230508_111707_create_census_tables;
 mod m20230511_104804_create_node;
 mod m20230511_104811_create_record;
 mod m20230511_104814_create_content;
@@ -8,6 +7,7 @@ mod m20230511_104823_create_client_info;
 mod m20230511_104830_create_content_audit;
 mod m20230511_104838_create_execution_metadata;
 mod m20230511_104937_create_key_value;
+mod m20230599_999999_create_census_tables;
 mod m20231107_004843_create_audit_stats;
 mod m20240213_190221_add_fourfours_stats;
 mod m20240322_205213_add_content_audit_index;
@@ -31,7 +31,7 @@ impl MigratorTrait for Migrator {
             Box::new(m20230511_104830_create_content_audit::Migration),
             Box::new(m20230511_104838_create_execution_metadata::Migration),
             Box::new(m20230511_104937_create_key_value::Migration),
-            Box::new(m20230508_111707_create_census_tables::Migration),
+            Box::new(m20230599_999999_create_census_tables::Migration),
             Box::new(m20231107_004843_create_audit_stats::Migration),
             Box::new(m20240213_190221_add_fourfours_stats::Migration),
             Box::new(m20240322_205213_add_content_audit_index::Migration),
